feat: move AltinnActivityExtensions to new package
Creates new package for telemetry related functionality in
ServiceDefaults called `Altinn.Authorization.ServiceDefaults.Telemetry`
and moves `AltinnActivityExtensions` (and implementation classes) to
this package. Also adds a TypeForwardedTo attribute to the new assembly
from the old one to maintain binary compatibility.
