+++ src/Altinn.Authorization.ServiceDefaults/src/ServiceDefaults.Telemetry/ActivityHelper.cs
