+++ src/Altinn.Authorization.ServiceDefaults/src/ServiceDefaults.Telemetry/Altinn.Authorization.ServiceDefaults.Telemetry.csproj
@@ -1,0 +1,7 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <ItemGroup>
+	  <InternalsVisibleTo Include="Altinn.Authorization.ServiceDefaults.Tests"/>
+  </ItemGroup>
+		
+</Project>
