+++ src/Altinn.Authorization.Utils.sln
@@ -75,6 +75,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Altinn.Cli.Jwks", "Altinn.C
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Altinn.Authorization.ServiceDefaults.Npgsql.TestSeed", "Altinn.Authorization.ServiceDefaults\src\ServiceDefaults.Npgsql.TestSeed\Altinn.Authorization.ServiceDefaults.Npgsql.TestSeed.csproj", "{261C441B-F4D7-4E25-A6CF-49316A19AB29}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Altinn.Authorization.ServiceDefaults.Telemetry", "Altinn.Authorization.ServiceDefaults\src\ServiceDefaults.Telemetry\Altinn.Authorization.ServiceDefaults.Telemetry.csproj", "{22B753AE-1913-4873-9139-76DBA3C9A3F5}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -165,6 +167,10 @@ Global
 		{261C441B-F4D7-4E25-A6CF-49316A19AB29}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{261C441B-F4D7-4E25-A6CF-49316A19AB29}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{261C441B-F4D7-4E25-A6CF-49316A19AB29}.Release|Any CPU.Build.0 = Release|Any CPU
+		{22B753AE-1913-4873-9139-76DBA3C9A3F5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{22B753AE-1913-4873-9139-76DBA3C9A3F5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{22B753AE-1913-4873-9139-76DBA3C9A3F5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{22B753AE-1913-4873-9139-76DBA3C9A3F5}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -201,6 +207,7 @@ Global
 		{A356DDAF-DD7B-4CCC-A069-C0EB37E40AEE} = {EC90677D-5F37-4A68-97A5-0FD840FCAD97}
 		{2D4570CD-39CF-45BE-9CCB-0327D97DD638} = {A356DDAF-DD7B-4CCC-A069-C0EB37E40AEE}
 		{261C441B-F4D7-4E25-A6CF-49316A19AB29} = {D3DEC557-4814-41FD-B74A-E5DCABCB724C}
+		{22B753AE-1913-4873-9139-76DBA3C9A3F5} = {D3DEC557-4814-41FD-B74A-E5DCABCB724C}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {3A2B25D1-5A10-4899-A920-AB7F206BE9D4}
